+++ android/src/main/kotlin/com/ipification/plugin/AuthenticateListener.kt
@@ -1,9 +1,8 @@
-package com.ipification.sdk.ip_sdk
-
+package com.ipification.plugin
 
 
 interface AuthenticationListener {
     fun onSuccess(result: String)
     fun onFail(result: AuthenticationError)
-    fun onError(result: AuthenticationError)
+    fun onIMCancel()
 }