+++ android/src/main/kotlin/com/ipification/plugin/AuthenticationError.kt
@@ -1,4 +1,4 @@
-package com.ipification.sdk.ip_sdk
+package com.ipification.plugin
 
 class AuthenticationError(var error_code: ErrorCode = ErrorCode.AUTHENTICATE_FAIL, var error_message:String?="") {
 }